requirements.txt
# Core
Django==5.2.4
python-dotenv==1.0.0
decouple==3.8
djangorestframework==3.14.0

# Database
psycopg2-binary==2.9.9

# Authentication
djangorestframework-simplejwt==5.3.1

# CORS
django-cors-headers==4.3.1

# Email
sendgrid==6.11.0
python-http-client==3.3.7

# Development
pylint-django==2.5.3
pylint==3.0.3

# Testing
pytest==7.4.2
pytest-django==4.7.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0
